requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
Jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
joblib==1.5.3
pandas==2.3.3
scikit-learn==1.8.0
PyYAML==6.0.3
numpy==2.3.5
scipy==1.16.3
python-dateutil==2.9.0.post0
pytz==2025.2
